NMS-8801: Refactored and reused AnyServerX509TrustManager into core/lib.
