+++ core/jmx/api/src/main/java/org/opennms/netmgt/jmx/connection/PasswordStrategy.java
@@ -37,6 +37,7 @@ import javax.net.ssl.SSLSocketFactory;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.TrustManagerFactory;
 
+import org.opennms.core.utils.AnyServerX509TrustManager;
 import org.slf4j.LoggerFactory;
 
 public interface PasswordStrategy {
