+++ features/ticketing/tsrm-integration/pom.xml
@@ -69,6 +69,10 @@
 	</build>
 	<dependencies>
 		<dependency>
+			<groupId>org.opennms.core</groupId>
+			<artifactId>org.opennms.core.lib</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>org.opennms.features.ticketing</groupId>
 			<artifactId>org.opennms.features.ticketing.api</artifactId>
 		</dependency>
